+++ lib/contacts/add_friend_page/add_friend_page.dart
@@ -5,7 +5,6 @@ import 'package:uplink/contacts/add_friend_page/models/models_export.dart';
 import 'package:uplink/contacts/add_friend_page/widgets/widgets_export.dart';
 import 'package:uplink/l10n/main_app_strings.dart';
 import 'package:uplink/utils/mock/data/data_export.dart';
-import 'package:uplink/utils/mock/models/models_export.dart';
 import 'package:uplink/utils/utils_export.dart';
 
 class AddFriendPage extends StatefulWidget {
