+++ lib/contacts/user_profile_page/models/user_about.dart
@@ -10,40 +10,37 @@ class UserAbout extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16),
-      child: Column(
-        children: [
-          Row(
-            children: const [
-              UIcon(
-                UIcons.about,
-                color: UColors.textMed,
-              ),
-              SizedBox.square(
-                dimension: 8,
-              ),
-              UText(
-                UAppStrings.profileIndexPage_aboutTitle,
-                textStyle: UTextStyle.H3_tertiaryHeader,
-              ),
-            ],
-          ),
-          const SizedBox.square(
-            dimension: 18,
-          ),
-          Align(
-            alignment: Alignment.centerLeft,
-            child: UText(
-              (user.isBlocked == true) || (user.about == null)
-                  ? '-'
-                  : user.about!,
-              textStyle: UTextStyle.B1_body,
-              textColor: UColors.white,
+    return Column(
+      children: [
+        Row(
+          children: const [
+            UIcon(
+              UIcons.about,
+              color: UColors.textMed,
             ),
+            SizedBox.square(
+              dimension: 8,
+            ),
+            UText(
+              UAppStrings.profileIndexPage_aboutTitle,
+              textStyle: UTextStyle.H3_tertiaryHeader,
+            ),
+          ],
+        ),
+        const SizedBox.square(
+          dimension: 18,
+        ),
+        Align(
+          alignment: Alignment.centerLeft,
+          child: UText(
+            (user.isBlocked == true) || (user.about == null)
+                ? '-'
+                : user.about!,
+            textStyle: UTextStyle.B1_body,
+            textColor: UColors.white,
           ),
-        ],
-      ),
+        ),
+      ],
     );
   }
 }
