+++ lib/contacts/user_profile_page/models/user_note.dart
@@ -6,56 +6,53 @@ class UserNote extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16),
-      child: Column(
-        children: [
-          Row(
-            children: const [
-              UIcon(
-                UIcons.add_note,
-                color: UColors.textMed,
-              ),
-              SizedBox.square(
-                dimension: 8,
-              ),
-              UText(
-                'Note',
-                textStyle: UTextStyle.H3_tertiaryHeader,
-              ),
-            ],
-          ),
-          const SizedBox.square(
-            dimension: 16,
-          ),
-          TextField(
-            style: UTextStyle.H5_fifthHeader.style.returnTextStyleType(
-              color: Colors.white,
+    return Column(
+      children: [
+        Row(
+          children: const [
+            UIcon(
+              UIcons.add_note,
+              color: UColors.textMed,
             ),
-            textAlignVertical: TextAlignVertical.top,
-            textInputAction: TextInputAction.done,
-            cursorColor: UColors.textDark,
-            autocorrect: false,
-            onChanged: (value) {},
-            decoration: InputDecoration(
-              filled: true,
-              fillColor: UColors.foregroundDark,
-              border: OutlineInputBorder(
-                borderSide: BorderSide.none,
-                borderRadius: BorderRadius.circular(4),
-              ),
-              hintText: 'Tap to add note',
-              hintStyle: UTextStyle.H5_fifthHeader.style.returnTextStyleType(
-                fontStyle: FontStyle.italic,
-              ),
+            SizedBox.square(
+              dimension: 8,
             ),
-            maxLines: 3,
+            UText(
+              'Note',
+              textStyle: UTextStyle.H3_tertiaryHeader,
+            ),
+          ],
+        ),
+        const SizedBox.square(
+          dimension: 16,
+        ),
+        TextField(
+          style: UTextStyle.H5_fifthHeader.style.returnTextStyleType(
+            color: Colors.white,
           ),
-          const SizedBox.square(
-            dimension: 24,
+          textAlignVertical: TextAlignVertical.top,
+          textInputAction: TextInputAction.done,
+          cursorColor: UColors.textDark,
+          autocorrect: false,
+          onChanged: (value) {},
+          decoration: InputDecoration(
+            filled: true,
+            fillColor: UColors.foregroundDark,
+            border: OutlineInputBorder(
+              borderSide: BorderSide.none,
+              borderRadius: BorderRadius.circular(4),
+            ),
+            hintText: 'Tap to add note',
+            hintStyle: UTextStyle.H5_fifthHeader.style.returnTextStyleType(
+              fontStyle: FontStyle.italic,
+            ),
           ),
-        ],
-      ),
+          maxLines: 3,
+        ),
+        const SizedBox.square(
+          dimension: 24,
+        ),
+      ],
     );
   }
 }
