+++ lib/contacts/user_profile_page/user_profile_page.dart
@@ -1,8 +1,7 @@
-// ignore_for_file: prefer_const_constructors
-
 import 'package:flutter/material.dart';
 import 'package:ui_library/ui_library_export.dart';
 import 'package:uplink/contacts/user_profile_page/models/models_export.dart';
+import 'package:uplink/utils/mock/data/data_export.dart';
 import 'package:uplink/utils/mock/models/mock_contact.dart';
 
 class UserProfilePage extends StatefulWidget {
@@ -19,25 +18,18 @@ class UserProfilePage extends StatefulWidget {
 
 class _UserProfilePageState extends State<UserProfilePage> {
   late MockContact user;
-  bool hideInfo = false;
 
   @override
   void initState() {
     super.initState();
     user = widget.user;
-    if (user.isBlocked == true) {
-      hideInfo = true;
-    }
   }
 
   @override
   Widget build(BuildContext context) {
-    final _size = MediaQuery.of(context).size;
-
     return Scaffold(
       resizeToAvoidBottomInset: true,
       body: CustomScrollView(
-        shrinkWrap: true,
         slivers: [
           SliverToBoxAdapter(
             child: Stack(
@@ -51,39 +43,69 @@ class _UserProfilePageState extends State<UserProfilePage> {
                         dimension: 114,
                       ),
                       UserBasicInfo(user: user),
-                      const SizedBox.square(
-                        dimension: 16,
-                      ),
-                      SizedBox(
-                        width: _size.width - 32,
-                        child: CTAButton(
-                          user: user,
-                        ),
-                      ),
-                      const SizedBox.square(
-                        dimension: 16,
-                      ),
-                      UserGroupedInfo(user: user),
-                      const SizedBox.square(
-                        dimension: 24,
-                      ),
-                      const UDivider(),
-                      const SizedBox.square(
-                        dimension: 16,
-                      ),
-                      UserAbout(user: user),
-                      const SizedBox.square(
-                        dimension: 24,
-                      ),
-                      // TODO(yijing): add note function
-                      UserNote(),
-                      UserLinkedAccounts(),
                     ],
                   ),
                 ),
               ],
             ),
           ),
+          SliverPadding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            sliver: SliverList(
+              delegate: SliverChildListDelegate([
+                const SizedBox.square(
+                  dimension: 16,
+                ),
+                CTAButton(
+                  user: user,
+                ),
+                const SizedBox.square(
+                  dimension: 16,
+                ),
+                UserGroupedInfo(user: user),
+                const SizedBox.square(
+                  dimension: 24,
+                ),
+              ]),
+            ),
+          ),
+          const SliverToBoxAdapter(
+            child: UDivider(height: 1),
+          ),
+          SliverPadding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            sliver: SliverList(
+              delegate: SliverChildListDelegate([
+                const SizedBox.square(
+                  dimension: 16,
+                ),
+                UserAbout(user: user),
+                const SizedBox.square(
+                  dimension: 24,
+                ),
+                // TODO(yijing): add note function
+                const UserNote(),
+              ]),
+            ),
+          ),
+          if (linkedAccountsList.isNotEmpty && user.isBlocked == false)
+            const SliverToBoxAdapter(
+              child: UDivider(height: 1),
+            ),
+          if (user.isBlocked == false)
+            SliverList(
+              delegate: SliverChildBuilderDelegate(
+                (context, index) {
+                  return UserLinkedAccount(
+                    networkUsername:
+                        linkedAccountsList[index]['username'] as String,
+                    networkProfile:
+                        linkedAccountsList[index]['network'] as NetworkProfiles,
+                  );
+                },
+                childCount: linkedAccountsList.length,
+              ),
+            )
         ],
       ),
     );
